Moved preloaded_models module into composite_models sub-package
