+++ halotools/empirical_models/__init__.py
@@ -5,7 +5,6 @@ from __future__ import (absolute_import, division, print_function,
 
 from .mock_factories import *
 from .model_factories import *
-from .preloaded_models import *
 from .model_defaults import *
 from .abunmatch import *
 from .model_helpers import *
@@ -15,3 +14,4 @@ from .phase_space_models import *
 from .sfr_models import *
 from .occupation_models import *
 from .smhm_models import *
+from .composite_models import *
