+++ halotools/empirical_models/composite_models/__init__.py
@@ -1,0 +1,1 @@
+from .preloaded_models import *
