+++ halotools/empirical_models/composite_models/preloaded_models.py
@@ -9,18 +9,18 @@ from __future__ import (
 
 import numpy as np
 
-from . import model_factories, model_defaults
-from .occupation_models import hod_components as hoc
-from .occupation_models import zheng07_components
-from .occupation_models import leauthaud11_components 
-from .occupation_models import tinker13_components 
+from .. import model_factories, model_defaults
+from ..occupation_models import hod_components as hoc
+from ..occupation_models import zheng07_components
+from ..occupation_models import leauthaud11_components 
+from ..occupation_models import tinker13_components 
 
-from .smhm_models import smhm_components
-from .sfr_models import BinaryGalpropInterpolModel
-from .phase_space_models import NFWPhaseSpace, TrivialPhaseSpace
-from .abunmatch import ConditionalAbunMatch
+from ..smhm_models import smhm_components
+from ..sfr_models import BinaryGalpropInterpolModel
+from ..phase_space_models import NFWPhaseSpace, TrivialPhaseSpace
+from ..abunmatch import ConditionalAbunMatch
 
-from ..sim_manager import FakeMock, FakeSim, sim_defaults
+from ...sim_manager import FakeMock, FakeSim, sim_defaults
 
 
 __all__ = ['Zheng07', 'SmHmBinarySFR', 'Leauthaud11', 'Campbell15', 'Hearin15', 'Tinker13']
