+++ halotools/empirical_models/composite_models/tests/test_preloaded_models.py
