+++ halotools/empirical_models/occupation_models/tests/test_leauthaud11_hod.py
@@ -5,7 +5,7 @@ from copy import copy
 
 from .. import Leauthaud11Cens, Leauthaud11Sats
 
-from ...preloaded_models import Leauthaud11
+from ...composite_models import Leauthaud11
 from ... import model_defaults
